Refactor the message handler out of bot to allow for easier testing
