+++ src/bot.js
@@ -3,13 +3,9 @@ const winston = require("winston");
 const _ = require("lodash");
 const admin = require("firebase-admin");
 require("./api");
-const { whoPlaysCommand } = require("./commands/whoPlays");
-const { iPlayCommand } = require("./commands/iPlay");
-const { gamesCommand } = require("./commands/games");
-const { whatsNewCommand } = require("./commands/whatsNew");
-const { witnessMeCommand } = require("./commands/witnessMe");
 const { extractGuildId } = require("./discordUtils");
 const { getWatchedUsers, addPlayer, ALREADY_PLAYS } = require("./store/store");
+const { handleMessage } = require("./bot/messageHandler");
 
 const discordAuth =
   process.env.NODE_ENV == "production" ? {} : require("../auth.json");
@@ -40,71 +36,17 @@ bot.on("ready", event => {
   logger.info(`${bot.username} - (${bot.id})`);
 });
 
-const allCommands = [
-  whoPlaysCommand,
-  iPlayCommand,
-  gamesCommand,
-  whatsNewCommand,
-  witnessMeCommand
-];
-
-const playBotId = "456628305911873536";
-const botMentionMatcher = /<@456628305911873536> ?/;
-
-const isTestChannel = channel => channel == "457011209372303371";
-const isTestUser = user => user == "232704012070158336";
-
-bot.on("message", (user, userID, channelID, originalMessage, event) => {
-  const enableTestCommands = isTestChannel(channelID);
-  const guildId = extractGuildId(event);
-
-  if (userID == playBotId) {
-    logger.debug("Message was from play-bot himself");
-    return;
-  }
-
-  // if (!isTestChannel(channelID)) {
-  //   logger.debug("Only test channel right now");
-  //   return;
-  // }
-
-  const commands = allCommands.filter(
-    i => i.test == false || enableTestCommands
-  );
-
-  if (!originalMessage.match(botMentionMatcher)) {
-    logger.debug("Message was not a command");
-    return;
-  }
-
-  const message = originalMessage.replace(botMentionMatcher, "");
-  const matchingCommand = commands.find(currentCommand =>
-    currentCommand.matches(message)
-  );
-  if (matchingCommand) {
-    matchingCommand.handle({
-      db,
-      user,
-      userId: userID,
-      channelId: channelID,
-      guildId,
-      message,
-      bot,
-      event,
-      logger
-    });
-  }
-  if (message.match(/^commands/i)) {
-    bot.sendMessage({
-      to: channelID,
-      message:
-        "Try asking me:\n" +
-        commands
-          .map((i, index) => `[${index + 1}]: <@${playBotId}> ${i.example}`)
-          .join("\n")
-    });
-  }
-});
+bot.on("message", (user, userId, channelId, originalMessage, event) =>
+  handleMessage({
+    user,
+    userId,
+    channelId,
+    originalMessage,
+    event,
+    bot,
+    db
+  })
+);
 
 bot.on("presence", (user, userID, status, game, event) => {
   // if (!isTestUser(userID)) {
