+++ src/bot/info.js
@@ -1,0 +1,7 @@
+const playBotId = "456628305911873536";
+const botMentionMatcher = /<@456628305911873536> ?/;
+
+module.exports = {
+  botId: playBotId,
+  botMentionMatcher: botMentionMatcher
+};
